+++ lib/widgets/charts/empty_dashboards_widget.dart
@@ -6,7 +6,7 @@ import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/widgets/app_bar/dashboards_app_bar.dart';
+import 'package:lotti/widgets/dashboards/dashboards_app_bar.dart';
 import 'package:lottie/lottie.dart';
 import 'package:url_launcher/url_launcher.dart';