+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- Upgraded dependencies
+- Refactoring: BloC for state management on dashboards page
+
+### Added:
+- Filter dashboards by categories
+- Tests for dashboards page
+
+## [0.8.343] - 2023-05-04
 ### Fixed:
 - Margin below habit completion dialog
 - Awaiting in getIt init